# backend/api-document/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Metrade API",
    "description": "This API is used for managing users, products, orders, order items, carts, cart items, token.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+123456789"
          },
          "balance": {
            "type": "number",
            "example": 1000
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "seller"
            ],
            "example": "user"
          },
          "photo_url": {
            "type": "string",
            "example": "https://example.com/images/user.png"
          },
          "is_verified": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "banned",
              "deleted"
            ],
            "example": "active"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Laptop"
          },
          "description": {
            "type": "string",
            "example": "A powerful laptop for developers."
          },
          "price": {
            "type": "number",
            "example": 1200
          },
          "stock_quantity": {
            "type": "number",
            "example": 10
          },
          "pickup_point": {
            "type": "string",
            "enum": [
              "Myllypuro",
              "Karamalmi",
              "Myyrmäki"
            ],
            "example": "Myllypuro"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "active",
              "sold"
            ],
            "example": "active"
          }
        }
      },
...
    },
    "securitySchemes": {
      "accessTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    }
  },
  "security": [
    {
      "accessTokenCookie": [],
      "refreshTokenCookie": []
    }
  ],
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all users",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update user status",
        "operationId": "updateUserStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User status updated"
          }
        }
      }
    },
    "/admin/product": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all products",
        "operationId": "adminGetAllProducts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/product/activate/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Activate a product",
        "operationId": "activateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product activated successfully"
          }
        }
      }
    },
    "/admin/product/delete/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a product",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          }
        }
      }
    }
  }
}